YARN-3300. Outstanding_resource_requests table should not be shown in AHS. Contributed by Xuan Gong
